alteracoes

+++ fileManager/lib/fileManager.pm
@@ -25,7 +25,7 @@ our %EXPORT_TAGS = ( 'all' => [ qw(
 	
 ) ] );
 
-our @EXPORT_OK = qw( registraParametros filtraString contaOcorrencias converteDataHora filtraDataHora atualizaRegistro geraArqLista );
+our @EXPORT_OK = qw( registraParametros filtraString contaOcorrencias converteDataHora filtraDataHora atualizaRegistro filtraTamanho geraArqLista );
 
 our @EXPORT = qw(
 	
@@ -34,15 +34,21 @@ our @EXPORT = qw(
 #Armazena a primeira e a segunda linha do arquivo de registro (que contem a opção e a string) a ser passada pelo gerenciador C++
 sub registraParametros {
 
-    open(my $entradaReg, "<", $_[0]) or die "Erro! O arquivo não pode ser aberto: $!";
+    open(my $entrParam, "<", $_[0]) or die "Erro! O arquivo $_[0] não pode ser aberto: $!";
 
-    my @linhasReg = <$entradaReg>;
-    chomp($linhasReg [0]);
-    chomp($linhasReg [1]);
-    my @opcao = split (/::/, $linhasReg [0]);   #pega o parametro da primeira linha
-    my @strDsj = split (/::/, $linhasReg [1]);  #pega o parametro da segunda linha
+    my @linhasParam = <$entrParam>;
 
-    return $opcao [1], $strDsj [1];
+    #extrai os parametros das três linhas do parametros.txt
+    my $opcao = $linhasParam [0];  
+    my $strDsj = $linhasParam [1]; 
+    my $arqReg = $linhasParam [2];
+
+    chomp($opcao);
+    chomp($strDsj);
+
+    close $entrParam or die "$entrParam: $!";
+
+    return $opcao, $strDsj, $arqReg;
 }
 
 #filtra nomes de arquivo por string passada no arquivo de registros
@@ -52,7 +58,6 @@ sub filtraString{
 
     my $ocorrencia = 0;
     my @lista;
-    my $ocorreu = 0;
 
     open(my $entradaReg, "<", $nomeArqReg) or die "Erro! O arquivo não pode ser aberto: $!";
 
@@ -65,21 +70,17 @@ sub filtraString{
     while (<$entradaReg>) {
 
         chomp($_);
+        my @sptLinha = split(/::/, $_); #separa as cacterísticas nominais dos arquivos
 
-        my @sptLinha = split(/::/, $_); #separa identificador e nome: $splitLinha[0] -> identificador e $splitLinha[1] -> nome
-
-        #toda vez que encontrar identificador diferente de ';P' (ou seja, a partir da terceira linha do arq de reg)
-        if($sptLinha [0] ne ";P"){
-            #verifica se a string desejada está contida no nome do arquivo e armazena o mesmo numa lista
-            if($sptLinha [0] =~ /$strDsj/){
-                $ocorrencia++;
-                $lista[$ocorrencia]{nome} = $sptLinha [0];
-                $lista[$ocorrencia]{diretorio} = $sptLinha [1];
-                $lista[$ocorrencia]{dataHora} = $sptLinha [2];
-                $lista[$ocorrencia]{tamanho} = $sptLinha [3];
-            }
+        #verifica se a string desejada está contida no nome do arquivo e armazena o mesmo numa lista
+        if($sptLinha [0] =~ /$strDsj/){
+            $ocorrencia++;
+            $lista[$ocorrencia]{nome} = $sptLinha [0];
+            $lista[$ocorrencia]{diretorio} = $sptLinha [1];
+            $lista[$ocorrencia]{dataHora} = $sptLinha [2];
+            $lista[$ocorrencia]{tamanho} = $sptLinha [3];
         }
-
+        
     }
 
     close $entradaReg or die "$entradaReg: $!";
@@ -112,34 +113,31 @@ sub contaOcorrencias {
         my @sptLinha = split(/::/, $_);
 
         #quando encontrar um nome de arquivo abre esse arquivo
-        if($sptLinha [0] ne ";P"){
             
-            my $ocorrenciaArq = 0; #ocorrencia em cada arquivo
-            my $numLinOcorr = 0;    #numero de linhas em que aparecem a string
+        my $ocorrenciaArq = 0; #ocorrencia em cada arquivo
+        my $numLinOcorr = 0;    #numero de linhas em que aparecem a string
             
-            open(my $entradaArq, "<", $sptLinha [0]) or die "Erro! O arquivo não pode ser aberto: $!";
+        open(my $entradaArq, "<", $sptLinha [0]) or die "Erro! O arquivo não pode ser aberto: $!";
 
-            #percorre todas as linhas do arquivo aberto e conta em quantas linhas há ocorrencia e quantas ocorrencias naquele arquivo
-            while (<$entradaArq>) {
-                if ($_ =~ /$strDsj/) {
-                    $numLinOcorr++; 
-                }
-                while ($_ =~ m/($strDsj)/g){
-                    $ocorrenciaArq++;
-                }
+        #percorre todas as linhas do arquivo aberto e conta em quantas linhas há ocorrencia e quantas ocorrencias naquele arquivo
+        while (<$entradaArq>) {
+            if ($_ =~ /$strDsj/) {
+                $numLinOcorr++; 
             }
-            $ocorrenciaTotal += $ocorrenciaArq;
-
-            close $entradaArq or die "$entradaArq: $!";
-            
-            #salva numa lista todas ocorencias de cada arquivo e em quantas linhas se deram
-            if ($ocorrenciaArq != 0) {
-                $lista[$numLinhasArqReg]{nome} = $sptLinha[0];
-                $lista[$numLinhasArqReg]{ocorrenciaArq} = $ocorrenciaArq;
-                $lista[$numLinhasArqReg]{numLinOcorr} = $numLinOcorr;
-                $numLinhasArqReg++;
+            while ($_ =~ m/($strDsj)/g){
+                $ocorrenciaArq++;
             }
+        }
+        $ocorrenciaTotal += $ocorrenciaArq;
 
+        close $entradaArq or die "$entradaArq: $!";
+            
+        #salva numa lista todas ocorencias de cada arquivo e em quantas linhas se deram
+        if ($ocorrenciaArq != 0) {
+            $lista[$numLinhasArqReg]{nome} = $sptLinha[0];
+            $lista[$numLinhasArqReg]{ocorrenciaArq} = $ocorrenciaArq;
+            $lista[$numLinhasArqReg]{numLinOcorr} = $numLinOcorr;
+            $numLinhasArqReg++;
         }
 
     }
@@ -187,12 +185,6 @@ sub filtraDataHora {
 
     open(my $entradaReg, "<", $arqReg) or die "Erro! O arquivo não pode ser aberto: $!";
 
-    my $i=0;
-    while(<$entradaReg>){
-        $i++;
-        if($i==2){last;}
-    }
-
     while (<$entradaReg>) {
         chomp($_);
         
@@ -227,23 +219,23 @@ sub atualizaRegistro{
                     chomp($_);
                     my @sptLinha = split(/::/, $_);
                     #quando encontrar um nome de arquivo abre esse arquivo
-                    if($sptLinha [0] ne ";P"){                      
-                        open(my $entradaArq, "<", $sptLinha [0]) or die "Erro! O arquivo não pode ser aberto: $!";
-                            my $statusArq = stat ($entradaArq);
-                            my $tamanho = $statusArq->size;
-                            
-                            if ($tamanho >= 1024) {
-                                $tamanho /= 1024;
-                                $tamanho = int ($tamanho * 100) / 100;
-                                $tamanho = $tamanho." KB";
-                            }
-                            elsif ($tamanho >= (1024 * 1024)) {
-                                $tamanho /= (1024 * 1024);
-                                $tamanho = int ($tamanho * 100) / 100;
-                                $tamanho = $tamanho." MB";
-                            }
-                            else {$tamanho = $tamanho." bytes";}
-                            my $dataHora = ctime ($statusArq->mtime);
+                                         
+                    open(my $entradaArq, "<", $sptLinha [0]) or die "Erro! O arquivo não pode ser aberto: $!";
+                        my $statusArq = stat ($entradaArq);
+                        my $tamanho = $statusArq->size;
+                        
+                        if ($tamanho >= 1024) {
+                            $tamanho /= 1024;
+                            $tamanho = int ($tamanho * 100) / 100;
+                            $tamanho = $tamanho." KB";
+                        }
+                        elsif ($tamanho >= (1024 * 1024)) {
+                            $tamanho /= (1024 * 1024);
+                            $tamanho = int ($tamanho * 100) / 100;
+                           $tamanho = $tamanho." MB";
+                        }
+                        else {$tamanho = $tamanho." bytes";}
+                        my $dataHora = ctime ($statusArq->mtime);
 
                             #depois tenta isso, tem que ter o modulo File::Slurp
                             #my $strSubs = read_file "registro.txt";
@@ -251,11 +243,10 @@ sub atualizaRegistro{
                             #$strSubs =~ s/$sptLinha[3]/$tamanho/g;
                             #write_file "registro.txt", $strSubs;
 
-                            open(my $arqTemp,">>", 'temporario.txt');
-                                print $arqTemp $sptLinha[0]."::".$sptLinha[1]."::".ctime($statusArq->mtime)."::".$tamanho."\n";
-                            close $arqTemp;
-                        close $entradaArq or die "$entradaArq: $!";
-                    }
+                        open(my $arqTemp,">>", 'temporario.txt');
+                            print $arqTemp $sptLinha[0]."::".$sptLinha[1]."::".ctime($statusArq->mtime)."::".$tamanho."\n";
+                        close $arqTemp;
+                    close $entradaArq or die "$entradaArq: $!";
 
                 }
                 
@@ -263,18 +254,80 @@ sub atualizaRegistro{
     rename 'temporario.txt', 'registro.txt';
 }
 
+sub filtraTamanho {
+
+    (my $arqReg, my $faixaTamDsj) = @_;
+
+    my @lista;
+    my $ocorrencia = 0;
+
+    my @tamDsj = split(/-->/,$faixaTamDsj);
+    my @tamI = split(/\s/, $tamDsj [0]);
+    my @tamF = split(/\s/, $tamDsj [1]);
+
+    if ($tamI [1] eq "KB") {
+        $tamI [0] = int($tamI[0]) * 1024;
+    }
+
+    if ($tamF [1] eq "KB") {
+        $tamF [0] = int($tamF[0]) * 1024;
+    }
+
+    $lista[$ocorrencia]{nome} = "Nome";
+    $lista[$ocorrencia]{diretorio} = "Diretorio";
+    $lista[$ocorrencia]{dataHora} = "Data e Hora";
+    $lista[$ocorrencia]{tamanho} = "Tamanho";
+
+    open(my $entradaReg, "<", $arqReg) or die "Erro! O arquivo não pode ser aberto: $!";
+
+    while (<$entradaReg>) {
+        chomp($_);
+        
+        my @sptLinha = split (/::/, $_);
+        
+        my @tamArq = split(/\s/, $sptLinha[3]);
+
+        if ($tamArq [1] eq "KB") {
+            $tamArq [0] = ($tamArq [0]) * 1024;
+        }
+        
+        if ($tamI [0] <= $tamArq [0] && $tamArq [0] <= $tamF [0]) {
+            $ocorrencia++;
+            $lista[$ocorrencia]{nome} = $sptLinha[0];
+            $lista[$ocorrencia]{diretorio} = $sptLinha[1];
+            $lista[$ocorrencia]{dataHora} = $sptLinha[2];
+            $lista[$ocorrencia]{tamanho} = $sptLinha[3];
+        }
+    }
+
+    close $entradaReg or die "$entradaReg: $!";
+
+    return $ocorrencia, @lista;
+
+}
+
 #gera arquivo contendo uma lista com as saídas desejadas (filtragem) de acordo com cada opção passada no arquivo de registro
 sub geraArqLista {
 
-    my $tabela = $_[0];
+    (my $ocorrencia, my @lista) = @_;
+    my $tabela = Text::TabularDisplay->new;
+    if ($ocorrencia == 0) {
+        print ("\n\n*** NENHUM RESULTADO ENCONTRADO! ***\n");
+    }
+    else {
+        print("\nArquivos encontrados: \n"); 
+        for my $href ( @lista ) {
+            $tabela->add($href->{nome}, $href->{diretorio}, $href->{dataHora}, $href->{tamanho});
+        }
 
-    unlink "lista-de-retorno.txt";
-    open(my $entrada, ">>", "lista-de-retorno.txt") or die "Erro! O arquivo não pode ser gerado: $!";
-    
-    print $entrada $tabela->render;
-    print $tabela->render;
+        unlink "lista-de-retorno.txt";
+        open(my $entrada, ">>", "lista-de-retorno.txt") or die "Erro! O arquivo não pode ser gerado: $!";
+        
+        print $entrada $tabela->render;
+        print $tabela->render;
 
-    close $entrada or die "$entrada: $!";
+        close $entrada or die "$entrada: $!";
+    }
 
 }
